refactor: move open-api-doc to src
