+++ api/src/certification/session-management/application/session-live-alert-route.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
 
-import { responseObjectErrorDoc } from '../../../../src/shared/infrastructure/open-api-doc/response-object-error-doc.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
+import { responseObjectErrorDoc } from '../../../shared/infrastructure/open-api-doc/response-object-error-doc.js';
 import { assessmentSupervisorAuthorization } from '../../shared/application/pre-handlers/session-supervisor-authorization.js';
 import { sessionLiveAlertController } from './session-live-alert-controller.js';
