+++ api/src/shared/infrastructure/open-api-doc/authentication/response-authentication-doc.js
