+++ api/src/shared/infrastructure/open-api-doc/pole-emploi/erreur-doc.js
