+++ api/tests/unit/infrastructure/open-api-doc/pole-emploi/envois-doc_test.js
@@ -1,4 +1,4 @@
-import { poleEmploiEnvoisDoc } from '../../../../../lib/infrastructure/open-api-doc/pole-emploi/envois-doc.js';
+import { poleEmploiEnvoisDoc } from '../../../../../src/shared/infrastructure/open-api-doc/pole-emploi/envois-doc.js';
 import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Open API Doc | Pole Emploi | Envois Documentation', function () {
